refactor that does not change output but makes dropping in the client code easier
