+++ bia-export/bia_export/cli.py
@@ -4,7 +4,9 @@ import json
 from rich.logging import RichHandler
 from typing_extensions import Annotated
 from pathlib import Path
-from .website_conversion import create_studies, create_ec_images
+from .website_export.studies import create_study
+from .website_export.images import create_ec_images
+from .website_export.website_models import StudyCreationContext
 from typing import List
 import json
 
@@ -16,11 +18,10 @@ logger = logging.getLogger()
 app = typer.Typer()
 
 
+
 @app.command()
 def website_study(
-    accession_id_list: Annotated[
-        List[str], typer.Argument(help="Accession IDs of the studies to export")
-    ],
+    id_list: Annotated[List[str], typer.Argument(help="IDs of the studies to export")],
     root_directory: Annotated[
         Path,
         typer.Option(
@@ -38,8 +39,17 @@ def website_study(
     ] = Path("bia-images-export.json"),
 ):
 
-    abs_root = root_directory.resolve()
-    studies_map = create_studies(accession_id_list, abs_root)
+    if root_directory:
+        abs_root = root_directory.resolve()
+
+    studies_map = {}
+    for id in id_list:
+        if abs_root:
+            context = StudyCreationContext(root_directory=abs_root, accession_id=id)
+        else:
+            context = StudyCreationContext(study_uuid=id)
+        study = create_study(context)
+        studies_map[study.accession_id] = study.model_dump(mode="json")
 
     logging.info(f"Writing study info to {output_filename.absolute()}")
     with open(output_filename, "w") as output:
